Store marker ID inside `ReplicatedArchetypes`
